backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sendgrid
pyjwt
python-dotenv
passlib[bcrypt]
pydantic[email]
